changed speed

+++ Constants.h
@@ -11,21 +11,22 @@ constexpr float BONUS_SPEED = 3.f;
 
 constexpr float CARRIAGE_SPEED = 7.f;
 
-constexpr int UNBREAKABLE_HP = 999;
+constexpr float UNBREAKABLE_HP = 999;
 
-constexpr int BALL_START_X = 400.f;
-constexpr int BALL_START_Y = 400.f;
+constexpr float BALL_START_X = 400.f;
+constexpr float BALL_START_Y = 400.f;
 
-constexpr int SLOW_BALL_SPEED = 2.f;
+constexpr float SLOW_BALL_SPEED = 2.f;
 
-constexpr int DEFEAT = 5;
+constexpr float DEFEAT = 5;
 
-constexpr int BIG_CARRIAGE_WIDTH = 180.f;
-constexpr int CARRIAGE_WIDTH = 120.f;
-constexpr int CARRIAGE_HEIGHT = 20.f;
+constexpr float BIG_CARRIAGE_WIDTH = 180.f;
+constexpr float CARRIAGE_WIDTH = 120.f;
+constexpr float CARRIAGE_HEIGHT = 20.f;
 
-constexpr int SHIELD_BOUNCE_Y = 560.f;
+constexpr float SHIELD_BOUNCE_Y = 560.f;
 
-constexpr int BLOCK_WIDTH = 60.f;
-constexpr int BLOCK_HEIGHT = 25.f;
+constexpr float BLOCK_WIDTH = 60.f;
+constexpr float BLOCK_HEIGHT = 25.f;
+
